chore(webapp): Reorganised code in subfolders
