+++ quizmaster-client-app/src/components/Common/Contestant.ts
