+++ quizmaster-client-app/src/components/QuizHosting/QuizPlaying/QuestionMarker.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { cleanup } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
-import HostLobby from "./HostLobby";
+import HostLobby from "../QuizCreating/HostLobby";
 import ReactDOM from "react-dom";
 
 jest.mock("react-router-dom", () => ({
@@ -18,6 +18,6 @@ describe("When testing directly", () => {
   afterEach(cleanup);
   it("renders without crashing", () => {
     const div = document.createElement("div");
-    ReactDOM.render(<HostLobby />, div);
+    ReactDOM.render(<HostLobby refreshContestants={true} />, div);
   });
 });
