+++ quizmaster-client-app/src/components/QuizHosting/QuizPlaying/QuestionMarker.tsx
