+++ quizmaster-client-app/src/components/QuizHosting/QuizPlaying/QuizQuestionDisplay.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
-import "./HostLobby.css";
+import "../QuizCreating/HostLobby.css";
 import { Paper, LinearProgress } from "@material-ui/core";
 
 const useStyles = makeStyles((theme) => ({
